requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
distro==1.9.0
Django==5.0.4
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.15.1
drf-yasg==1.21.7
fake-useragent==1.4.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
future==0.18.3
google==3.0.0
Google-Search-API @ git+https://github.com/abenassi/Google-Search-API@546a59cc22d3260c60a2faf9afe6477168ead627
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
oauthlib==3.2.2
openai==1.11.1
openapi-codec==1.3.2
packaging==23.2
pillow==10.3.0
pluggy==1.4.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
pytest==7.3.1
pytest-mock==3.10.0
pytest-print==0.3.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.0
selenium==2.53.6
simplejson==3.19.2
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.11.0
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
vcrpy==6.0.1
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
